app/requirements.txt
pinecone-io
pinecone-client
openai
langchain==0.0.237
streamlit-chat
streamlit
transformers
llama_index

tqdm
pytesseract
Pillow

click
flask
msal
pathvalidate
pyyaml
requests_oauthlib

python-dotenv
chromadb

weaviate-client==3.22.1


click
flask
msal
pathvalidate
pyyaml
requests_oauthlib
